lxd/storage/block: Move GetDisksByID to prevent import cycle
GetDisksByID is anyway solely used inside lxd/storage/block so moving it makes most sense.
